feat: extract AvatarCropModal to shared component, add crop to onboarding
AvatarCropModal moved from a local function in Profile.jsx to
src/components/ui/AvatarCropModal.jsx so it can be shared.

PhotoBioStep now opens the crop modal when a photo is picked — same
drag-to-reposition, scroll-to-zoom experience as the /profile photo editor.
Profile.jsx imports from the shared component instead of defining inline.
